Extract a parent class AddTaskCommand
Some children classes of Command have the same behaviour in the method
execute, except for how the task is created.

Let's extract out the execute method into a new parent class named
AddTaskCommand.

Using inheritance is preferable over composition in this situation
because the common behaviors are not composable.
