+++ src/main/java/luna/command/DeadlineCommand.java
@@ -2,14 +2,13 @@ package luna.command;
 
 import java.time.LocalDate;
 
-import luna.storage.Storage;
 import luna.task.Deadline;
-import luna.task.TaskList;
+import luna.task.Task;
 
 /**
  * Represents the {@code deadline} command.
  */
-public class DeadlineCommand extends IntermediateCommand {
+public class DeadlineCommand extends AddTaskCommand {
     private final String name;
     private final LocalDate deadline;
 
@@ -19,11 +18,7 @@ public class DeadlineCommand extends IntermediateCommand {
     }
 
     @Override
-    public String execute(TaskList taskList, Storage<TaskList> storage) {
-        Deadline newTask = new Deadline(name, deadline);
-        taskList.add(newTask);
-        saveTaskList(taskList, storage);
-        return "Got it. I've added this task:\n  " + newTask + "\nNow you have " + taskList.getSize()
-                + " tasks in the list.";
+    protected Task getTaskToAdd() {
+        return new Deadline(name, deadline);
     }
 }
