+++ src/main/java/luna/command/EventCommand.java
@@ -2,14 +2,13 @@ package luna.command;
 
 import java.time.LocalDate;
 
-import luna.storage.Storage;
 import luna.task.Event;
-import luna.task.TaskList;
+import luna.task.Task;
 
 /**
  * Represents the {@code event} command.
  */
-public class EventCommand extends IntermediateCommand {
+public class EventCommand extends AddTaskCommand {
     private final String name;
     private final LocalDate start;
     private final LocalDate end;
@@ -21,11 +20,7 @@ public class EventCommand extends IntermediateCommand {
     }
 
     @Override
-    public String execute(TaskList taskList, Storage<TaskList> storage) {
-        Event newTask = new Event(name, start, end);
-        taskList.add(newTask);
-        saveTaskList(taskList, storage);
-        return "Got it. I've added this task:\n  " + newTask + "\nNow you have " + taskList.getSize()
-                + " tasks in the list.";
+    protected Task getTaskToAdd() {
+        return new Event(name, start, end);
     }
 }
