+++ src/main/java/luna/command/TodoCommand.java
@@ -1,13 +1,12 @@
 package luna.command;
 
-import luna.storage.Storage;
-import luna.task.TaskList;
+import luna.task.Task;
 import luna.task.ToDo;
 
 /**
  * Represents the {@code todo} command.
  */
-public class TodoCommand extends IntermediateCommand {
+public class TodoCommand extends AddTaskCommand {
     private final String name;
 
     public TodoCommand(String name) {
@@ -15,11 +14,7 @@ public class TodoCommand extends IntermediateCommand {
     }
 
     @Override
-    public String execute(TaskList taskList, Storage<TaskList> storage) {
-        ToDo newTask = new ToDo(name);
-        taskList.add(newTask);
-        saveTaskList(taskList, storage);
-        return "Got it. I've added this task:\n  " + newTask + "\nNow you have " + taskList.getSize()
-                + " tasks in the list.";
+    protected Task getTaskToAdd() {
+        return new ToDo(name);
     }
 }
